added refresh token

+++ controllers/user.js
@@ -1,6 +1,10 @@
 const User = require('../models/user');
 const Community = require('../models/community');
-const { BadRequestError, NotFoundError } = require('../errors');
+const {
+  BadRequestError,
+  NotFoundError,
+  UnauthenticatedError,
+} = require('../errors');
 const { StatusCodes } = require('http-status-codes');
 const NotificationTokenSchema = require('../models/notification');
 
@@ -104,11 +108,28 @@ const removeCommunity = async (req, res) => {
 };
 
 const LogoutUser = async (req, res) => {
+  const cookie = req?.cookies?.jwt;
+  if (!cookie) {
+    throw new UnauthenticatedError('Unauthorized access');
+  }
+
+  const user = await User.findOneAndUpdate(
+    { refreshToken: cookie },
+    { refreshToken: '' },
+    { new: true }
+  );
+
+  if (!user) {
+    res.clearCookie('jwt', { httpOnly: true, maxAge: 24 * 60 * 60 * 1000 });
+    throw new NotFoundError('User not found');
+  }
+
   await NotificationTokenSchema.findOneAndDelete({
     userId: req.user.userId,
   });
 
-  res.status(StatusCodes.OK).json({ message: 'Token deleted for user' });
+  res.clearCookie('jwt', { httpOnly: true, maxAge: 24 * 60 * 60 * 1000 });
+  res.status(StatusCodes.OK).json({ message: 'Logout successfully' });
 };
 
 module.exports = {
